+++ src/main/java/net/silentchaos512/gems/world/GemsWorldFeatures.java
@@ -206,15 +206,22 @@ public final class GemsWorldFeatures {
     }
 
     private static void addChaosOre(Biome biome, Random random) {
-        int count = MathUtils.nextIntInclusive(random, 1, 2);
-        int size = MathUtils.nextIntInclusive(random, 12, 18);
-        int maxHeight = MathUtils.nextIntInclusive(random, 15, 25);
-        //SilentGems.LOGGER.debug("    Biome {}: add chaos ore (size {}, count {}, maxHeight {})", biome, size, count, maxHeight);
-        addOre(biome, MiscOres.CHAOS.asBlock(), size, count, 5, maxHeight, d -> true);
+        int minCount = GemsConfig.COMMON.worldGenChaosOreMinCount.get();
+        int maxCount = GemsConfig.COMMON.worldGenChaosOreMaxCount.get();
+        if (minCount > maxCount) {
+            minCount = maxCount;
+        }
+        if (maxCount > 0) {
+            int count = MathUtils.nextIntInclusive(random, minCount, maxCount);
+            int size = MathUtils.nextIntInclusive(random, 12, 18);
+            int maxHeight = MathUtils.nextIntInclusive(random, 15, 25);
+            //SilentGems.LOGGER.debug("    Biome {}: add chaos ore (size {}, count {}, maxHeight {})", biome, size, count, maxHeight);
+            addOre(biome, MiscOres.CHAOS.asBlock(), size, count, 5, maxHeight, d -> true);
+        }
     }
 
     private static void addEnderOre(Biome biome, Random random) {
-        addOre(biome, MiscOres.ENDER.asBlock(), 32, 1, 10, 70, state -> state.getBlock() == Blocks.END_STONE, d -> true);
+        addOre(biome, MiscOres.ENDER.asBlock(), 32, GemsConfig.COMMON.worldGenEnderOreCount.get(), 10, 70, state -> state.getBlock() == Blocks.END_STONE, d -> true);
     }
 
     private static void addSilverOre(Biome biome, Random random) {